refactor(category): update hook imports
